app/requirements.txt
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
psycopg2==2.8.4
requests==2.22.0
